pyproject.toml
[tool.poetry]
name = "praxis-core"
version = "0.3.0"
description = "Local autonomous assistant for PRAXIS-1"
authors = ["Cliff May <you@example.com>"]
packages = [{ include = "praxis" }]

[tool.poetry.dependencies]
python = "^3.11"
click = "^8.1"
requests = "^2.32"
python-dotenv = "^1.0"
pydantic = "^2.7"
ollama = "*"
openai = ">=1.14"
pyautogui = "^0.9"
pydantic-settings = "^2.9.1"

[tool.poetry.scripts]
praxis-cli = "praxis.cli:main"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"

